feat(api): migrated address service

+++ packages/api/src/modules/identity/application/use-cases/user/permanently-delete-user.use-case.ts
@@ -3,14 +3,14 @@ import type { Database } from '@yukinu/db/drizzle'
 import { TRPCError } from '@trpc/server'
 import { utapi } from '@yukinu/uploadthing'
 
-import type { UserIdActionDto } from '@/modules/identity/application/dtos/user/user-id-action.dto'
 import type { UserRepository } from '@/modules/identity/domain/repositories/user.repository'
+import type { OneUserDto } from '@/modules/identity/types'
 
 import { AbstractUseCase } from '@/shared/abstracts/abstract.use-case'
 
 export class PermanentlyDeleteUserUseCase extends AbstractUseCase<
-  UserIdActionDto.Input,
-  UserIdActionDto.Output
+  OneUserDto.Input,
+  OneUserDto.Output
 > {
   constructor(
     private readonly _db: Database,
@@ -19,19 +19,23 @@ export class PermanentlyDeleteUserUseCase extends AbstractUseCase<
     super()
   }
 
-  public async execute(
-    input: UserIdActionDto.Input,
-  ): Promise<UserIdActionDto.Output> {
-    const user = await this._userRepo.find(input)
+  public async execute(input: OneUserDto.Input): Promise<OneUserDto.Output> {
+    const [user] = await this._userRepo.find([input], {}, { limit: 1 })
     if (!user)
       throw new TRPCError({
         code: 'NOT_FOUND',
         message: `User with id ${input.id} not found`,
       })
 
-    await this._userRepo.delete(input)
+    if (user.deletedAt === null)
+      throw new TRPCError({
+        code: 'BAD_REQUEST',
+        message: `User with id ${input.id} is not deleted`,
+      })
+
+    await this._userRepo.delete([input])
     if (user.image) await utapi.deleteFiles(user.image.split('/').pop() ?? '')
 
-    return { id: input.id }
+    return user
   }
 }
